Move UI related code to celestia/win32
